+++ tests/test_builtin_alloc.rs
@@ -1,10 +1,53 @@
 #![cfg(feature = "bundled_jit_alloc")]
 
-use closure_ffi::{BareFn, BareFnMut, BareFnOnce};
+#[allow(unused_imports)]
+use closure_ffi::{traits::FnPtr, BareFn, BareFnMut, BareFnOnce};
+
+#[test]
+fn test_infer_closure() {
+    let bare_closure: BareFn<'_, unsafe extern "C" fn(usize) -> u32>;
+    bare_closure = BareFn::new(|arg| arg as _);
+
+    assert_eq!(unsafe { bare_closure.bare().call((42,)) }, 42);
+}
+
+#[test]
+fn test_boxed_fn() {
+    let boxed_dyn = Box::new(|n: usize| 2 * n) as Box<dyn Fn(usize) -> usize>;
+    let bare_closure = BareFn::new_c(boxed_dyn);
+
+    assert_eq!(unsafe { bare_closure.bare().call((42,)) }, 84);
+}
+
+#[test]
+fn test_ref_dyn() {
+    let closure = |n: usize| 2 * n;
+    let bare_closure = BareFn::new_c(&closure as &dyn Fn(usize) -> usize);
+
+    assert_eq!(unsafe { bare_closure.bare().call((42,)) }, 84);
+}
+
+#[test]
+fn test_mut_dyn() {
+    let mut sum = 0;
+    let mut closure = |n: usize| sum += n;
+
+    let bare_closure = BareFnMut::new_c(&mut closure as &mut dyn FnMut(usize));
+    let bare = bare_closure.bare();
+
+    unsafe {
+        bare(5);
+        bare(3);
+        bare(7);
+    }
+
+    drop(bare_closure);
+    assert_eq!(sum, 15);
+}
 
 #[test]
 fn test_stateless_fn() {
-    let bare_closure = BareFn::new_c(move |n: usize| 2 * n);
+    let bare_closure = BareFn::new_c(|n: usize| 2 * n);
 
     let bare = bare_closure.bare();
     assert_eq!(unsafe { bare(5) }, 10);
@@ -13,7 +56,7 @@ fn test_stateless_fn() {
 #[test]
 fn test_borrow_fn() {
     let array = [0, 5, 10, 15, 20];
-    let bare_closure = BareFn::new_c(|n: usize| array[n]);
+    let bare_closure = BareFn::new_c(|n| array[n]);
 
     let bare = bare_closure.bare();
     assert_eq!(unsafe { bare(3) }, 15);